fix: three refusals of programs that were fine
- an integer literal above i64::MAX was a lex error, so no U64 constant
  past 2^63 (FNV's offset basis, splitmix's increment) could be written
- a lone tuple of names was not seen as covering its type
- a recursive function that also called a helper outside its group was
  treated as mutually recursive, and its measure was not inferred

Also: `ty_of` knows the numeric result of a saturated prelude call.

+++ src/ast.rs
@@ -38,7 +38,7 @@ impl Ty {
 pub enum Pat {
     Wild,
     Var(String),
-    Int(i64),
+    Int(i128),
     Float(f64),
     Str(String),
     Char(char),
@@ -56,7 +56,7 @@ pub struct Expr {
 
 #[derive(Clone, Debug)]
 pub enum ExprKind {
-    Int(i64),
+    Int(i128),
     Float(f64),
     Str(String),
     Char(char),
